+++ src/apify_helpers.js
@@ -5,6 +5,9 @@ const { APIFY_API_ENDPOINTS, DEFAULT_KEY_VALUE_STORE_KEYS, LEGACY_PHANTOM_JS_CRA
     DEFAULT_ACTOR_MEMORY_MBYTES } = require('./consts');
 const { wrapRequestWithRetries } = require('./request_helpers');
 
+// Key of field to use internally to compute changes in fields.
+const ACTOR_ID_REFERENCE_FIELD_KEY = 'referenceActorId';
+
 const createDatasetUrls = (datasetId, cleanParamName) => {
     const createDatasetUrl = (format) => {
         return `${APIFY_API_ENDPOINTS.datasets}/${datasetId}/items?${cleanParamName}=true&attachment=true&format=${format}`;
@@ -119,8 +122,8 @@ const enrichActorRun = async (z, run, storeKeysToInclude = []) => {
 
     // Attach Apify app URL to detail of run
     run.detailsPageUrl = run.actorTaskId
-        ? `https://my.apify.com/tasks/${run.actorTaskId}#/runs/${run.id}`
-        : `https://my.apify.com/actors/${run.actId}#/runs/${run.id}`;
+        ? `https://console.apify.com/actors/tasks/${run.actorTaskId}/runs/${run.id}`
+        : `https://console.apify.com/actors/${run.actId}/runs/${run.id}`;
 
     // Omit fields, which are useless for Zapier users.
     return _.omit(run, OMIT_ACTOR_RUN_FIELDS);
@@ -195,13 +198,13 @@ const getOrCreateKeyValueStore = async (z, storeIdOrName) => {
 };
 
 /**
- * It pickes from input schema prefill values.
+ * It picks from input schema prefill values.
  * NOTE: Input schema was validated on app, we don't have to check structure here.
- * @param inputSchemaStringJSON
+ * @param inputSchema
  */
-const getPrefilledValuesFromInputSchema = (inputSchemaStringJSON) => {
+const getPrefilledValuesFromInputSchema = (inputSchema) => {
     const prefilledObject = {};
-    const { properties } = JSON.parse(inputSchemaStringJSON);
+    const { properties } = inputSchema;
 
     Object.keys(properties).forEach((propKey) => {
         if (properties[propKey].prefill) prefilledObject[propKey] = properties[propKey].prefill;
@@ -214,81 +217,231 @@ const getPrefilledValuesFromInputSchema = (inputSchemaStringJSON) => {
 };
 
 /**
+ * Prefix input field key with input string.
+ * @param fieldKey
+ * @returns string
+ */
+const prefixInputFieldKey = (fieldKey) => {
+    return `input-${fieldKey}`;
+};
+
+/**
+ * Parse input field key to get original key.
+ * @param fieldKey
+ * @returns string
+ */
+const parseInputFieldKey = (fieldKey) => {
+    return fieldKey.replace('input-', '');
+};
+
+/**
+ * Converts Apify input schema to Zapier input fields.
+ * Input schema spec.
+ * https://docs.apify.com/platform/actors/development/actor-definition/input-schema Fields schema
+ * spec. https://zapier.github.io/zapier-platform-schema/build/schema.html#fieldschema
+ * @param inputSchema
+ * @param actor
+ */
+const createFieldsFromInputSchemaV1 = (inputSchema, actor) => {
+    const { properties, required, description } = inputSchema;
+    const fields = [
+        // The first fies is info box with input schema description or actor title, same as on Apify platform.
+        {
+            label: actor.title,
+            key: prefixInputFieldKey(`actor-${actor.id}-info`),
+            type: 'copy',
+            helpText: description || `${actor.title} Input, see [documentation](https://apify.com/${actor.username}/${actor.name}) `
+                + 'for detailed fields description.',
+        },
+    ];
+    // eslint-disable-next-line no-restricted-syntax
+    for (const [propertyKey, definition] of Object.entries(properties)) {
+        // eslint-disable-next-line no-continue
+        if (definition.editor === 'hidden') continue;
+        // NOTE: Handle sectionCaption with info box with helpText. It is not possible to do stackable fields in Zapier.
+        if (definition.sectionCaption && definition.sectionCaption.length) {
+            const helpText = definition.sectionDescription
+                ? `${definition.sectionCaption} - ${definition.sectionDescription}`
+                : definition.sectionCaption;
+            fields.push({
+                label: definition.sectionCaption,
+                key: prefixInputFieldKey(`sectionCaption-${propertyKey}`),
+                type: 'copy',
+                helpText,
+            });
+        }
+        const field = {
+            label: definition.title,
+            helpText: definition.description,
+            key: prefixInputFieldKey(propertyKey),
+            required: required && required.includes(propertyKey),
+            // NOTE: From Zapier docs: A default value that is saved the first time a Zap is created.
+            // It is what what prefill is in Apify input schema.
+            default: definition.prefill,
+            // NOTE: From Zapier docs: An example value that is not saved.
+            // It is what what default is in Apify input schema.
+            placeholder: definition.default,
+        };
+        switch (definition.type) {
+            case 'string': {
+                // NOTE: Cannot provide alternative in fields schema for options pattern, minLength, maxLength, nullable
+                // These options will not cover UI validation and we need to handle it in code.
+                field.type = 'string'; // editor = textfield, datepicker
+                if (['javascript', 'python'].includes(definition.editor)) {
+                    field.type = 'code';
+                } else if (definition.editor === 'textarea') {
+                    field.type = 'text';
+                } else if (definition.editor === 'datepicker') {
+                    field.type = 'datetime';
+                } else if (definition.editor === 'select') {
+                    field.choices = {};
+                    definition.enum.forEach((key, i) => {
+                        field.choices[key] = definition.enumTitles ? definition.enumTitles[i] : key;
+                    });
+                }
+                if (definition.isSecret) {
+                    field.type = 'password';
+                }
+                break;
+            }
+            case 'integer': {
+                // NOTE: Cannot provide alternative in fields schema for options maximum, minimum, unit, nullable
+                field.type = 'integer';
+                break;
+            }
+            case 'boolean':
+                // NOTE: Cannot provide alternative in fields schema for options groupCaption, groupDescription, nullable
+                field.type = 'boolean';
+                break;
+            case 'array': {
+                const parsedPrefillValue = definition.prefill;
+                const parsedDefaultValue = definition.default;
+                // NOTE: Cannot provide alternative in fields schema for options placeholderKey, placeholderValue, patternKey,
+                // patternValue, maxItems, minItems, uniqueItems, nullable
+                if (definition.editor === 'json' || definition.editor === 'keyValue') {
+                    field.type = 'text';
+                    if (parsedPrefillValue) field.default = JSON.stringify(parsedPrefillValue, null, 2);
+                    else if (parsedDefaultValue) field.placeholder = JSON.stringify(parsedDefaultValue, null, 2);
+                } else if (['requestListSources', 'pseudoUrls', 'globs', 'stringList'].includes(definition.editor)) {
+                    // NOTE: These options are not supported in Zapier and Apify UI specific.
+                    // We will use stringList type instead for simplicity. We will covert them into spec. format before run.
+                    field.type = 'string';
+                    field.list = true;
+                    // NOTE: List can have just one default value, so pick just first one.
+                    if (parsedPrefillValue && Array.isArray(parsedPrefillValue) && parsedPrefillValue[0]) {
+                        const firstItem = parsedPrefillValue[0];
+                        if (typeof firstItem === 'string') field.default = firstItem;
+                        else if (typeof firstItem === 'object') field.default = firstItem.url || firstItem.purl || firstItem.glob;
+                        else field.default = firstItem; // NOTE: We do not know what it is, let's print it as it is, but it should not happen.
+                        field.placeholder = undefined;
+                    } else if (parsedDefaultValue && Array.isArray(parsedDefaultValue) && parsedDefaultValue[0]) {
+                        const firstItem = parsedDefaultValue[0];
+                        if (typeof firstItem === 'string') field.placeholder = firstItem;
+                        else if (typeof firstItem === 'object') field.placeholder = firstItem.url || firstItem.purl || firstItem.glob;
+                        else field.placeholder = firstItem; // NOTE: We do not know what it is, let's print it as it is, but it should not happen.
+                        field.default = undefined;
+                    } else {
+                        field.default = undefined;
+                        field.placeholder = undefined;
+                    }
+                }
+                break;
+            }
+            case 'object': {
+                if (definition.editor === 'json') {
+                    field.type = 'text';
+                } else if (definition.editor === 'proxy') {
+                    // This field is Apify specific, we do not support nice UI for it. Let's print note about it into UI.
+                    fields.push({
+                        label: 'Proxy',
+                        key: prefixInputFieldKey('proxyWarning'),
+                        type: 'copy',
+                        helpText: `${definition.title} depends on Apify platform and is not compatible with Zapier integration. `
+                            + 'We suggest setting this value in the Apify console',
+                    });
+                    field.type = 'text';
+                }
+                if (definition.prefill) {
+                    field.default = JSON.stringify(definition.prefill, null, 2);
+                } else if (field.default) {
+                    field.placeholder = JSON.stringify(definition.default, null, 2);
+                }
+                break;
+            }
+            default: {
+                // This should not happen.
+                console.log(`Unknown input schema type: ${definition.type}`, definition);
+                // eslint-disable-next-line no-continue
+                continue;
+            }
+        }
+        fields.push(field);
+    }
+    return fields;
+};
+
+const maybeGetInputSchemaFromActor = async (z, actor, buildTag) => {
+    const defaultBuild = actor.taggedBuilds && actor.taggedBuilds[buildTag];
+    if (defaultBuild) {
+        const buildResponse = await wrapRequestWithRetries(z.request, {
+            url: `${APIFY_API_ENDPOINTS.actors}/${actor.id}/builds/${defaultBuild.buildId}`,
+        });
+        const inputSchemaJSON = buildResponse.data && buildResponse.data.inputSchema;
+        try {
+            return JSON.parse(inputSchemaJSON);
+        } catch (err) {
+            // This should never happen, but if it does, we will ignore it
+            // and continue without input schema.
+        }
+    }
+};
+
+/**
  * This method loads additional input fields regarding actor default values.
  */
 const getActorAdditionalFields = async (z, bundle) => {
     const { actorId } = bundle.inputData;
-    if (!actorId) return [];
+    if (!actorId) return []; // Actor not selected yet, no additional fields to load.
+
+    const previousActorId = bundle.inputData[ACTOR_ID_REFERENCE_FIELD_KEY];
+    const wasActorChanged = actorId !== previousActorId; // If Actor ID changed from the last input field generation.
 
     const actorResponse = await wrapRequestWithRetries(z.request, {
         url: `${APIFY_API_ENDPOINTS.actors}/${actorId}`,
     });
 
     const actor = actorResponse.data;
-    const { build, timeoutSecs, memoryMbytes } = actor.defaultRunOptions;
-    const defaultActorBuildTag = build || BUILD_TAG_LATEST;
+    const { build: defaultBuild, timeoutSecs, memoryMbytes } = actor.defaultRunOptions;
+    const actorBuildTag = wasActorChanged
+        ? defaultBuild || BUILD_TAG_LATEST
+        : bundle.inputData.build || defaultBuild || BUILD_TAG_LATEST;
 
     let inputBody;
     let inputContentType;
-    let inputSchema;
     // Get input schema from build
-    const defaultBuild = actor.taggedBuilds && actor.taggedBuilds[defaultActorBuildTag];
-    if (defaultBuild) {
-        const buildResponse = await wrapRequestWithRetries(z.request, {
-            url: `${APIFY_API_ENDPOINTS.actors}/${actorId}/builds/${defaultBuild.buildId}`,
-        });
-        inputSchema = buildResponse.data && buildResponse.data.inputSchema;
-        if (inputSchema) {
-            inputContentType = 'application/json; charset=utf-8';
-            inputBody = JSON.stringify(getPrefilledValuesFromInputSchema(inputSchema), null, 2);
-        }
+    const inputSchema = await maybeGetInputSchemaFromActor(z, actor, actorBuildTag);
+    if (inputSchema) {
+        inputContentType = 'application/json; charset=utf-8';
+        inputBody = JSON.stringify(getPrefilledValuesFromInputSchema(inputSchema), null, 2);
     }
 
-    // Parse and stringify json input body if there is
-    if (actor.exampleRunInput && !inputSchema) {
-        const { body, contentType } = actor.exampleRunInput;
-        inputContentType = contentType;
-        // Try to parse JSON body
-        if (contentType.includes('application/json')) {
-            try {
-                const parsedBody = JSON.parse(body);
-                inputBody = JSON.stringify(parsedBody, null, 2);
-            } catch (err) {
-                // There can be invalid JSON, but show must go on.
-                inputBody = body;
-            }
-        }
-    }
-
-    let inputBodyHelpText = 'Input configuration for the actor.';
-    if (actor.isPublic) {
-        inputBodyHelpText += ` See [documentation](https://apify.com/${actor.username}/${actor.name}?section=input-schema) `
-            + 'for detailed fields description.';
-    }
-
-    return [
+    const baseFields = [
         {
-            label: 'Input body',
-            helpText: inputBodyHelpText,
-            key: 'inputBody',
-            required: false,
-            default: inputBody || '',
-            type: 'text', // NICE TO HAVE: Input type 'file' regarding content type
-        },
-        {
-            label: 'Input content type',
-            helpText: 'Specifies the `Content-Type` for the actor input body, e.g. `application/json`.',
-            key: 'inputContentType',
-            required: false,
-            default: inputContentType || '',
+            label: 'Reference Actor ID',
+            key: ACTOR_ID_REFERENCE_FIELD_KEY,
+            helpText: 'ID of Actor the UI was generated for.',
             type: 'string',
+            default: actorId,
+            computed: true, // This field is hidden in UI, used for checking if actorId change between input fields generation.
         },
         {
             label: 'Build',
             helpText: 'Tag or number of the build that you want to run, e.g. `latest`, `beta` or `1.2.34`.',
             key: 'build',
             required: false,
-            default: defaultActorBuildTag,
+            default: actorBuildTag,
+            // NOTE: Change build value recomputes fields as input schema can change.
+            altersDynamicFields: true,
             type: 'string',
         },
         {
@@ -311,6 +464,70 @@ const getActorAdditionalFields = async (z, bundle) => {
             type: 'string',
         },
     ];
+
+    if (inputSchema && (inputSchema.schemaVersion === 1 || !inputSchema.schemaVersion)) {
+        const fieldsFromInputSchema = createFieldsFromInputSchemaV1(inputSchema, actor);
+        return [
+            ...fieldsFromInputSchema,
+            {
+                label: 'Options',
+                key: `actor-${actor.id}-options`,
+                type: 'copy',
+                helpText: 'Actor options, see [documentation](https://docs.apify.com/platform/actors/running/usage-and-resources)'
+                    + ' for detailed description.',
+            },
+            ...baseFields,
+        ];
+    }
+
+    // Parse and stringify json input body if there is
+    if (actor.exampleRunInput) {
+        const { body, contentType } = actor.exampleRunInput;
+        inputContentType = contentType;
+        // Try to parse JSON body
+        if (contentType.includes('application/json')) {
+            try {
+                const parsedBody = JSON.parse(body);
+                inputBody = JSON.stringify(parsedBody, null, 2);
+            } catch (err) {
+                // There can be invalid JSON, but show must go on.
+                inputBody = body;
+            }
+        } else {
+            inputBody = body;
+        }
+    }
+    let inputBodyHelpText = 'Input configuration for the actor.';
+    if (actor.isPublic) {
+        inputBodyHelpText += ` See [documentation](https://apify.com/${actor.username}/${actor.name}/input-schema) `
+                + 'for detailed fields description.';
+    }
+    return [
+        {
+            label: 'Input body',
+            helpText: inputBodyHelpText,
+            key: 'inputBody',
+            required: false,
+            default: inputBody || '',
+            type: 'text', // NICE TO HAVE: Input type 'file' regarding content type
+        },
+        {
+            label: 'Input content type',
+            helpText: 'Specifies the `Content-Type` for the actor input body, e.g. `application/json`.',
+            key: 'inputContentType',
+            required: false,
+            default: inputContentType || '',
+            type: 'string',
+        },
+        ...baseFields,
+    ];
+};
+
+const printPrettyActorOrTaskName = (actorOrTask) => {
+    const idLikeName = actorOrTask.username ? `${actorOrTask.username}/${actorOrTask.name}` : actorOrTask.name;
+    return actorOrTask.title
+        ? `${actorOrTask.title} (${idLikeName})`
+        : idLikeName;
 };
 
 module.exports = {
@@ -322,4 +539,9 @@ module.exports = {
     getDatasetItems,
     getActorAdditionalFields,
     getPrefilledValuesFromInputSchema,
+    createFieldsFromInputSchemaV1,
+    maybeGetInputSchemaFromActor,
+    printPrettyActorOrTaskName,
+    parseInputFieldKey,
+    prefixInputFieldKey,
 };